flask_app/requirements.txt
flask
uvicorn
psycopg2
psycopg2-binary
bcrypt
PyJWT


